third_party/WebKit/LayoutTests/fast/spatial-navigation/snav-div-in-anchor-and-img-crash.html

<!DOCTYPE html>
<span><a href="#" id="start"><div></div></a></span>
<img style="display:block;">

<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script>
    test(() => {
        assert_true(!!window.testRunner);
        testRunner.overridePreference("WebKitTabToLinksPreferenceKey", 1);
        testRunner.overridePreference("WebKitSpatialNavigationEnabled", 1);
        document.getElementById("start").focus();
        eventSender.keyDown("ArrowDown");
    }, "No crash or assertion failure");
</script>
